requirements.txt
selenium
tqdm
mysql-connector-python
fastapi[standard]
requests
